+++ noxfile.py
@@ -14,8 +14,9 @@
 
 from __future__ import absolute_import
 
-import nox
 import os
+
+import nox
 
 
 def _install_dev_packages(session):
@@ -49,13 +50,16 @@ def _install_dev_packages(session):
 
 
 def _install_test_dependencies(session):
-    session.install('mock')
+    session.install('mock==3.0.5')
     session.install('pytest==4.6.4')
+    # 842 - Unit tests failing on CI due to failed import for coverage
+    # Might have something to do with the CircleCI image
+    # session.install('pytest-cov')
     session.install('retrying')
     session.install('unittest2')
 
 
-@nox.session(python=['2.7', '3.4', '3.5', '3.6'])
+@nox.session(python=['2.7', '3.5', '3.6'])
 def unit(session):
     """Run the unit test suite."""
 
@@ -69,6 +73,13 @@ def unit(session):
     session.run(
         'py.test',
         '--quiet',
+        # '--cov=opencensus',
+        # '--cov=context',
+        # '--cov=contrib',
+        # '--cov-append',
+        # '--cov-config=.coveragerc',
+        # '--cov-report=',
+        # '--cov-fail-under=97',
         'tests/unit/',
         'context/',
         'contrib/',
@@ -126,6 +137,17 @@ def lint_setup_py(session):
     session.install('docutils', 'pygments')
     session.run(
         'python', 'setup.py', 'check', '--restructuredtext', '--strict')
+
+
+# @nox.session(python='3.6')
+# def cover(session):
+#     """Run the final coverage report.
+#     This outputs the coverage report aggregating coverage from the unit
+#     test runs (not system test runs), and then erases coverage data.
+#     """
+#     session.install('coverage', 'pytest-cov')
+#     session.run('coverage', 'report', '--show-missing', '--fail-under=100')
+#     session.run('coverage', 'erase')
 
 
 @nox.session(python='3.6')